Revert "agregado products dashboard"
