--- greenwave/src/app/banned/page.jsx
@@ -1,13 +1,0 @@
-const Banned = () => {
-  return (
-    <div>
-      <h1>Baneado</h1>
-      <p>
-        Lo sentimos, pero no puedes acceder a esta página porque has sido
-        baneado.
-      </p>
-    </div>
-  );
-};
-
-export default Banned;
