+++ greenwave/src/app/components/dashboard/products/products.module.css
@@ -1,111 +1,13 @@
-.container {
-  background-color: var(--bgSoft);
-  padding: 20px;
-  border-radius: 10px;
-  margin-top: 20px;
-}
+.container{
+    background-color: var(--bgSoft);
+    padding: 20px;
+    border-radius: 10px;
+    margin-top: 20px;
+  }
 
-.top {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.table {
-  width: 100%;
-}
-
-.table td {
-  padding: 10px;
-  text-align: center; 
-}
-
-.user {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.userImage {
-  border-radius: 50%;
-  object-fit: cover;
-}
-
-.buttons {
-  display: flex;
-  gap: 10px;
-}
-
-.button {
-  padding: 5px 10px;
-  border-radius: 5px;
-  color: var(--text);
-  border: none;
-  cursor: pointer;
-}
-
-.buttonBottom {
-  padding: 5px 10px;
-  border-radius: 5px;
-  color: var(--text);
-  border: 1px solid white;
-  cursor: pointer;
-  margin-right: 10px;
-  margin-left: 10px;
-  background-color: #011c2e;
-}
-
-.view {
-  background-color: teal;
-}
-
-.delete {
-  background-color: crimson;
-}
-
-.status,
-.rol {
-  border-radius: 40px;
-  color: var(--textDashboard);
-  text-align: center;
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-}
-
-.status.active {
-  background-color: teal;
-}
-
-.status.banned {
-  background-color: crimson;
-}
-
-.rol.admin {
-  background-color: orange;
-}
-
-.rol.client {
-  background-color: blue;
-}
-
-.search {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  background-color: #2e374a;
-  padding: 10px;
-  border-radius: 10px;
-}
-
-.input {
-  background-color: transparent;
-  border: none;
-  color: var(--textDashboard);
-}
-
-.button:disabled,
-.buttonBottom:disabled {
-  opacity: 0.5; /* Reduce la opacidad para indicar que el botón está deshabilitado */
-  cursor: not-allowed; /* Cambia el cursor a 'not-allowed' cuando el botón está deshabilitado */
-}
+  .top{
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+  }
+
