+++ greenwave/src/app/components/users/globalUsers.jsx
@@ -1,8 +1,31 @@
 "use client";
-import React, { createContext, useContext,useState } from "react";
+import React, { createContext, useContext, useEffect, useState } from "react";
+import axios from "axios";
+import { useSession } from "next-auth/react";
 export const GlobalUser = createContext();
 export const UserGlobal = ({ children }) => {
+  const { data: session } = useSession();
+  const userData = session?.user;
   const [user, setUser] = useState({});
+  const fetchData = async () => {
+    try {
+      const response = await axios.get(
+        `http://localhost:3001/users/${userData?.email}`
+      );
+      const { data } = response;
+
+      setUser(data);
+    } catch (error) {
+      console.error("Error fetching user data:", error);
+      // Handle the error, show a message to the user, or redirect as needed.
+    }
+  };
+
+  useEffect(() => {
+    if (session?.user) {
+      fetchData();
+    }
+  }, [session?.user]);
 
   return (
     <GlobalUser.Provider
